TINY-9601: Move DataTransfer modules and tests back to agar from dragster
