+++ modules/agar/src/main/ts/ephox/agar/datatransfer/DataTransferItem.ts
