+++ modules/agar/src/main/ts/ephox/agar/datatransfer/DataTransferItemList.ts
