+++ modules/agar/src/main/ts/ephox/agar/dragndrop/DndEvents.ts
@@ -1,5 +1,6 @@
-import { DataTransferMode } from '@ephox/dragster';
 import { SugarElement } from '@ephox/sugar';
+
+import { setProtectedMode, setReadOnlyMode, setReadWriteMode } from '../datatransfer/Mode';
 
 const createDndEvent = (name: string) => (win: Window, x: number, y: number, dataTransfer: DataTransfer): DragEvent => {
   const event: any = document.createEvent('CustomEvent');
@@ -31,11 +32,11 @@ const isDefaultPrevented = (evt: DragEvent): boolean => evt.defaultPrevented;
 
 const dispatchDndEvent = (event: DragEvent, target: SugarElement<Node>): DragEvent => {
   if (event.type === 'dragstart') {
-    DataTransferMode.setReadWriteMode(event.dataTransfer);
+    setReadWriteMode(event.dataTransfer);
   } else if (event.type === 'drop') {
-    DataTransferMode.setReadOnlyMode(event.dataTransfer);
+    setReadOnlyMode(event.dataTransfer);
   } else {
-    DataTransferMode.setProtectedMode(event.dataTransfer);
+    setProtectedMode(event.dataTransfer);
   }
 
   target.dom.dispatchEvent(event);
