+++ modules/agar/src/main/ts/ephox/agar/file/FileList.ts
