--- modules/dragster/src/main/ts/ephox/dragster/api/DataTransfer.ts
@@ -1,7 +1,0 @@
-import { createDataTransfer, cloneDataTransfer, getDragImage } from '../datatransfer/DataTransfer';
-
-export {
-  createDataTransfer,
-  cloneDataTransfer,
-  getDragImage
-};
