--- modules/dragster/src/main/ts/ephox/dragster/api/DataTransferMode.ts
@@ -1,10 +1,0 @@
-import { setReadWriteMode, setProtectedMode, setReadOnlyMode, isInReadWriteMode, isInProtectedMode, isInReadOnlyMode } from '../datatransfer/Mode';
-
-export {
-  setReadWriteMode,
-  setProtectedMode,
-  setReadOnlyMode,
-  isInReadWriteMode,
-  isInProtectedMode,
-  isInReadOnlyMode
-};
