--- modules/dragster/src/main/ts/ephox/dragster/api/FileList.ts
@@ -1,3 +1,0 @@
-import { createFileList } from '../file/FileList';
-
-export { createFileList };
