--- modules/dragster/src/main/ts/ephox/dragster/api/Files.ts
@@ -1,6 +1,0 @@
-import { createFile, createFileFromString } from '../file/Files';
-
-export {
-  createFile,
-  createFileFromString
-};
