+++ modules/dragster/src/main/ts/ephox/dragster/api/Main.ts
@@ -1,17 +1,9 @@
-import * as DataTransfer from './DataTransfer';
-import * as DataTransferMode from './DataTransferMode';
 import * as DragApis from './DragApis';
 import * as Dragger from './Dragger';
-import * as FileList from './FileList';
-import * as Files from './Files';
 import MouseDrag from './MouseDrag';
 
 export {
-  DataTransfer,
-  DataTransferMode,
   DragApis,
   Dragger,
-  FileList,
-  Files,
   MouseDrag
 };
